Formatting components into one class per file
